Fix price fetching, chart rendering, and Dashboard crash bugs
- Switch OHLCVRow from rust_decimal::Decimal to f64 (fixes JSON
  serializing as strings instead of numbers, breaking chart rendering)
- Fix Dashboard PortfolioSummary calling usePrices in a loop (Rules of
  Hooks violation causing React crashes when asset count changed)
- Add empty-string API key guard in fetch_prices command
- Fix CoinGecko Box::leak memory leak in ticker_to_coin_id
- Add schema migrations to unlock_db (not just create_db)
- Fix dead SQL branch for (None, Some(to_ts)) in get_prices query
- Add User-Agent header to both providers to prevent throttling
- Remove rust_decimal dependency

+++ src-tauri/src/models.rs
@@ -1,4 +1,3 @@
-use rust_decimal::Decimal;
 use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
@@ -44,11 +43,11 @@ pub struct OHLCVRow {
     pub id: Option<i64>,
     pub asset_id: String,
     pub ts: i64,
-    pub open: Option<Decimal>,
-    pub high: Option<Decimal>,
-    pub low: Option<Decimal>,
-    pub close: Decimal,
-    pub volume: Option<Decimal>,
+    pub open: Option<f64>,
+    pub high: Option<f64>,
+    pub low: Option<f64>,
+    pub close: f64,
+    pub volume: Option<f64>,
 }
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
